+++ aci/Cargo.toml
@@ -39,7 +39,10 @@ cose = { package = "tee-attestation-verification-cose", path = "../cose", defaul
 crypto = { package = "tee-attestation-verification-crypto", path = "../crypto", default-features = false }
 serde_json = "1"
 
-[dev-dependencies]
+[target.'cfg(not(target_family = "wasm"))'.dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+
+[target.'cfg(target_family = "wasm")'.dev-dependencies]
 wasm-bindgen-test = "0.3"
 
 [target.'cfg(target_family = "wasm")'.dependencies]
