+++ aci/src/lib.rs
@@ -44,6 +44,7 @@
 //! # }
 //! ```
 
+mod base64;
 mod didx509;
 #[cfg(target_family = "wasm")]
 pub mod ffi;
